requirements.txt
django-cors-headers
django
djangorestframework
requests
beautifulsoup4
lxml
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
concurrently